==> R/10_analysis_doc.Rmd <==
---
title: "Model analysis"
author: "Benjamin Mercier"
date: "`r paste('Last updated on', format(Sys.time(), '%d %B, %Y'))`"
output: flexdashboard::flex_dashboard
---

```{r, echo = FALSE, message=FALSE, warning=FALSE}
# Load needed libraries
suppressPackageStartupMessages(library(rstan))
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(tidybayes))
suppressPackageStartupMessages(library(brms))
suppressPackageStartupMessages(library(bayesplot))
suppressPackageStartupMessages(library(flexdashboard))
suppressPackageStartupMessages(library(knitr))
# Options to run the chains faster
rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())
```

# Model 1
Column 1
--------------------------------------------------

### Description of the model
<br>

Model 1 fits the flux between a prey (i) and its predator (j) only based on mass-action (biomass of prey and abundance of predators).
The scaling parameters $\alpha$ is general to every predator.

\begin{align}
F_{ij}^{real} &= \alpha * B_i * \frac{B_j}{M_j}
\end{align}

```{r table_1, echo = FALSE}
kable(summary(output_stan_model1)$summary, caption = "Stan model 1 output")
```

Column 2
--------------------------------------------------

### Plots

```{r load model1, echo = FALSE, include = FALSE}
output_brms_model1 <- readRDS("../data/clean/model_outputs/output_brms_model1.RDS")
output_stan_model1 <- readRDS("../data/clean/model_outputs/output_stan_model1.RDS")
```

#### **Brms**

```{r plotting_1-1, echo = FALSE, out.width = "100%"}
plot(output_brms_model1)
```

#### **Stan**

```{r plotting_1-2, echo = FALSE, out.width = "100%"}
traceplot(output_stan_model1)
```

# Model 2
Column 1
--------------------------------------------------
### Description of the model
<br>

This second model is a slight variant of the first, in which we assign a specific $\alpha$ to each predator. In doing so, we can isolate difference between predator, to check if some predator seems to be different from the whole.

\begin{align}
F_{ij}^{real} &= \alpha_{j} * B_i * \frac{B_j}{M_j}
\end{align}

```{r table_2, echo = FALSE}
kable(summary(output_stan_model2)$summary, caption = "Stan model 2 output")
```

Column 2
--------------------------------------------------
### Plots

```{r load model2, echo = FALSE, include = FALSE}
output_brms_model2 <- readRDS("../data/clean/model_outputs/output_brms_model2.RDS")
output_stan_model2 <- readRDS("../data/clean/model_outputs/output_stan_model2.RDS")
```

#### **Brms**

```{r plotting_2-1, echo = FALSE, out.width = "100%"}
plot(output_brms_model2)
```

#### **Stan**

```{r plotting_2-2, echo = FALSE, out.width = "100%"}
traceplot(output_stan_model2)
```
